Revert "Feature/eventlistener"
