--- .scrutinizer.yml
@@ -1,21 +1,0 @@
-checks:
-  php:
-    simplify_boolean_return: true
-    properties_in_camelcaps: true
-    parameters_in_camelcaps: true
-    overriding_parameter: true
-    no_short_method_names:
-      minimum: '3'
-    no_goto: true
-    naming_conventions:
-      local_variable: '^[a-z][a-zA-Z0-9]*$'
-      abstract_class_name: ^Abstract|Factory$
-      utility_class_name: 'Utils?$'
-      constant_name: '^[A-Z][A-Z0-9]*(?:_[A-Z0-9]+)*$'
-      property_name: '^[a-z][a-zA-Z0-9]*$'
-      method_name: '^(?:[a-z]|__)[a-zA-Z0-9]*$'
-      parameter_name: '^[a-z][a-zA-Z0-9]*$'
-      interface_name: '^[A-Z][a-zA-Z0-9]*Interface$'
-      type_name: '^[A-Z][a-zA-Z0-9]*$'
-      exception_name: '^[A-Z][a-zA-Z0-9]*Exception$'
-      isser_method_name: '^(?:is|has|should|may|supports)'
