--- src/VectorNetworkProject/TheMix/event/BlockEventListener.php
@@ -1,158 +1,0 @@
-<?php
-/**
- * Copyright (c) 2018 VectorNetworkProject. All rights reserved. MIT license.
- *
- * GitHub: https://github.com/VectorNetworkProject/TheMix
- * Website: https://www.vector-network.tk
- */
-
-namespace VectorNetworkProject\TheMix\event;
-
-use pocketmine\block\Block;
-use pocketmine\event\block\BlockBreakEvent;
-use pocketmine\event\block\BlockPlaceEvent;
-use pocketmine\event\Listener;
-use pocketmine\item\Item;
-use pocketmine\Server;
-use VectorNetworkProject\TheMix\game\corepvp\blue\BlueCoreManager;
-use VectorNetworkProject\TheMix\game\corepvp\blue\BlueTeamManager;
-use VectorNetworkProject\TheMix\game\corepvp\red\RedCoreManager;
-use VectorNetworkProject\TheMix\game\corepvp\red\RedTeamManager;
-use VectorNetworkProject\TheMix\game\DefaultConfig;
-use VectorNetworkProject\TheMix\game\event\game\BreakCoreEvent;
-use VectorNetworkProject\TheMix\task\BlockReGeneratorTask;
-use VectorNetworkProject\TheMix\TheMix;
-
-class BlockEventListener implements Listener
-{
-    /**
-     * @param BlockBreakEvent $event
-     */
-    public function onOreBreak(BlockBreakEvent $event): void
-    {
-        $block = $event->getBlock();
-        $inventory = $event->getPlayer()->getInventory();
-        if (DefaultConfig::isDev() || $block->getLevel()->getName() !== DefaultConfig::getStageLevelName() || GameEventListener::isFinish()) {
-            $event->setCancelled();
-
-            return;
-        }
-        if ($block->getToolType() !== $inventory->getItemInHand()->getBlockToolType() && !DefaultConfig::isDev()) {
-            $event->setCancelled();
-
-            return;
-        }
-        if (RedTeamManager::getListCount() < 1 || BlueTeamManager::getListCount() < 1) {
-            $event->getPlayer()->sendMessage('プレイヤーが足りないのでブロックの採掘が不可能です。');
-            $event->setCancelled();
-
-            return;
-        }
-        $event->setDrops([]);
-        switch ($block->getId()) {
-            case Block::MELON_BLOCK:
-                $inventory->addItem(Item::get(Item::MELON, 0, 16));
-                TheMix::getInstance()->getScheduler()->scheduleDelayedTask(new BlockReGeneratorTask($block), 10 * 20);
-                break;
-            case Block::WOOD:
-                $inventory->addItem(Item::get(Item::WOODEN_PLANKS, 0, 4));
-                TheMix::getInstance()->getScheduler()->scheduleDelayedTask(new BlockReGeneratorTask($block), 15 * 20);
-                break;
-            case Block::DIAMOND_ORE:
-                $inventory->addItem(Item::get(Item::DIAMOND));
-                TheMix::getInstance()->getScheduler()->scheduleDelayedTask(new BlockReGeneratorTask($block), 60 * 20);
-                break;
-            case Block::EMERALD_ORE:
-                $inventory->addItem(Item::get(Item::EMERALD, 0, mt_rand(1, 3)));
-                TheMix::getInstance()->getScheduler()->scheduleDelayedTask(new BlockReGeneratorTask($block), 60 * 20);
-                break;
-            case Block::COAL_ORE:
-                $inventory->addItem(Item::get(Item::COAL));
-                TheMix::getInstance()->getScheduler()->scheduleDelayedTask(new BlockReGeneratorTask($block), 15 * 20);
-                break;
-            case Block::IRON_ORE:
-                $inventory->addItem(Item::get(Item::IRON_INGOT));
-                TheMix::getInstance()->getScheduler()->scheduleDelayedTask(new BlockReGeneratorTask($block), 20 * 20);
-                break;
-            case Block::GOLD_ORE:
-                $inventory->addItem(Item::get(Item::GOLD_INGOT));
-                TheMix::getInstance()->getScheduler()->scheduleDelayedTask(new BlockReGeneratorTask($block), 30 * 20);
-                break;
-            default:
-                if ($event->getPlayer()->getLevel()->getName() === DefaultConfig::getStageLevelName() && DefaultConfig::isDev() === false) {
-                    $event->setCancelled();
-
-                    return;
-                }
-        }
-    }
-
-    /**
-     * @param BlockBreakEvent $event
-     *
-     * @throws \ReflectionException
-     */
-    public function onBlockBreak(BlockBreakEvent $event)
-    {
-        $player = $event->getPlayer();
-        $block = $event->getBlock();
-        if (GameEventListener::isFinish()) {
-            $event->setCancelled();
-
-            return;
-        }
-        if ($player->getLevel()->getName() === Server::getInstance()->getDefaultLevel()->getName()) {
-            if ($player->isOp()) {
-                return;
-            }
-            $event->setCancelled();
-        }
-        if (!DefaultConfig::isDev()) {
-            if (RedCoreManager::isCore($block)) {
-                $event->setCancelled();
-                if (RedTeamManager::isJoined($player)) {
-                    return;
-                }
-                $revent = new BreakCoreEvent($player, BreakCoreEvent::RED);
-                $revent->call();
-                if (!$revent->isCancelled()) {
-                    RedCoreManager::reduceHP($revent->getDamage(), $player);
-                }
-            } elseif (BlueCoreManager::isCore($block)) {
-                $event->setCancelled();
-                if (BlueTeamManager::isJoined($player)) {
-                    return;
-                }
-                $bevent = new BreakCoreEvent($player, BreakCoreEvent::BLUE);
-                $bevent->call();
-                if (!$bevent->isCancelled()) {
-                    BlueCoreManager::reduceHP($bevent->getDamage(), $player);
-                }
-            }
-        }
-    }
-
-    /**
-     * @param BlockPlaceEvent $event
-     */
-    public function onBlockPlace(BlockPlaceEvent $event)
-    {
-        $player = $event->getPlayer();
-        if (GameEventListener::isFinish()) {
-            $event->setCancelled();
-
-            return;
-        }
-        if ($player->getLevel()->getName() === Server::getInstance()->getDefaultLevel()->getName()) {
-            if ($player->isOp()) {
-                return;
-            }
-            $event->setCancelled();
-        } elseif ($player->getLevel()->getName() === DefaultConfig::getStageLevelName()) {
-            if (DefaultConfig::isDev()) {
-                return;
-            }
-            $event->setCancelled();
-        }
-    }
-}
