+++ src/VectorNetworkProject/TheMix/event/level/TheItemSpawnEvent.php
@@ -6,19 +6,19 @@
  * Website: https://www.vector-network.tk
  */
 
-namespace VectorNetworkProject\TheMix\event;
+namespace VectorNetworkProject\TheMix\event\level;
 
 use pocketmine\event\entity\ItemSpawnEvent;
 use pocketmine\event\Listener;
 use pocketmine\item\Item;
 use VectorNetworkProject\TheMix\game\DefaultConfig;
 
-class LevelEventListener implements Listener
+class TheItemSpawnEvent implements Listener
 {
     /**
      * @param ItemSpawnEvent $event
      */
-    public function onItemSpawn(ItemSpawnEvent $event)
+    public function event(ItemSpawnEvent $event)
     {
         if (DefaultConfig::isDev()) {
             return;
