+++ src/VectorNetworkProject/TheMix/game/corepvp/TeamManager.php
@@ -43,12 +43,6 @@ abstract class TeamManager
         }
     }
 
-    public static function resetTeam(): void
-    {
-        RedTeamManager::ClearList();
-        BlueTeamManager::ClearList();
-    }
-
     /**
      * @param Player $player
      */
