+++ src/VectorNetworkProject/TheMix/task/UpdateScoreboardTask.php
@@ -44,8 +44,6 @@ class UpdateScoreboardTask extends Task
     {
         if (!$this->player->isOnline()) {
             $this->getHandler()->cancel();
-
-            return;
         }
         $scoreboard = $this->scoreboard;
         $scoreboard->setLine($this->player, 0, '§7'.date('Y/m/d H:i:s'));
@@ -55,7 +53,6 @@ class UpdateScoreboardTask extends Task
         $scoreboard->setLine($this->player, 6, 'Needed XP: §b'.LevelAPI::NeededXP($this->player));
         $scoreboard->setLine($this->player, 7, 'Gold: §6'.GoldAPI::getGold($this->player).'g');
         $scoreboard->setLine($this->player, 9, 'Streak: §c'.Streak::getStreak($this->player));
-        $scoreboard->setLine($this->player, 10, 'Event: none');
-        $scoreboard->setLine($this->player, 12, '§ewww.vector-network.tk  ');
+        $scoreboard->setLine($this->player, 11, '§ewww.vector-network.tk  ');
     }
 }
